% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{genderNameQueryCheck}
\alias{genderNameQueryCheck}
\title{Gender Name Query Check}
\usage{
genderNameQueryCheck()
}
\value{
Message that says how many queries are remaining today. The limit
is 1000 queries in a 24-hour period.
}
\description{
Gender Name Query Check
}
\examples{

# Gives number of checks remaining in the day
genderNameQueryCheck()

# If you run genderName again, it increments the number of
# available checks down by 1.
genderName("tom")
genderNameQueryCheck()

}

